apps/python-director/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
websockets>=12.0
pydantic>=2.5.3
python-multipart>=0.0.6
langchain>=0.1.0
langchain-groq>=0.0.1
groq>=0.4.2
python-dotenv>=1.0.1
fal-client>=0.3.3
httpx>=0.27.0
faiss-cpu
sentence-transformers
langchain-google-genai
langchain-anthropic
tiktoken
langchain-openai
huggingface_hub
Pillow
psutil
